docs(release): design system packages (#1316)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @rijkshuisstijl-community/web-components@1.1.0

### Minor Changes

-   8562d81: JSX now auto supports custom elemenents

### Patch Changes

-   651160d: Removed deprecated global file