fix: add custom condition for source mapping (#1935)
No longer tie the source mapping to just being in a dev environment as
this would trigger for users of the package in there own developer
environment causing references to non bundled source files.

Now it will only resolve the source files if you provide the custom
rhc:source condition to the build tool of the consuming package.

---------

Co-authored-by: Fionn Mac an Bhaird <fionn.mac@rijksoverheid.nl>
Co-authored-by: Casper Ravenhorst <casper@sinedub.io>